backend/pyproject.toml
[project]
name = "todo-skills-backend"
version = "0.1.0"
description = "Skills Library for Todo Full-Stack Web Application"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
    { name = "Todo App Team" }
]

dependencies = [
    # Web Framework
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",

    # Data Validation & ORM
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "sqlmodel>=0.0.14",

    # Database
    "psycopg2-binary>=2.9.9",
    "asyncpg>=0.29.0",

    # Authentication
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "argon2-cffi>=23.1.0",

    # Email validation
    "email-validator>=2.1.0",

    # Observability
    "structlog>=24.1.0",
    "prometheus-client>=0.19.0",

    # Utilities
    "python-dotenv>=1.0.0",
    "httpx>=0.26.0",

    # Database Migrations
    "alembic>=1.13.0",
]

[project.optional-dependencies]
dev = [
    # Testing
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "httpx>=0.26.0",

    # Linting & Formatting
    "ruff>=0.2.0",
    "black>=24.1.0",
    "mypy>=1.8.0",

    # Type stubs
    "types-passlib>=1.7.7",
    "types-python-jose>=3.3.4",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

# =============================================================================
# pytest configuration (T003)
# =============================================================================
[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
addopts = [
    "-v",
    "--tb=short",
    "--strict-markers",
    "-ra",
]
markers = [
    "unit: Unit tests (fast, isolated)",
    "integration: Integration tests (may use database)",
    "contract: Contract tests (schema validation)",
    "slow: Slow tests (AI operations, etc.)",
]
filterwarnings = [
    "ignore::DeprecationWarning",
]

# =============================================================================
# Coverage configuration
# =============================================================================
[tool.coverage.run]
source = ["src"]
branch = true
omit = [
    "*/tests/*",
    "*/__init__.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
    "if __name__ == .__main__.:",
]
fail_under = 80
show_missing = true

# =============================================================================
# Ruff configuration (T004) - Linting
# =============================================================================
[tool.ruff]
target-version = "py311"
line-length = 100
exclude = [
    ".git",
    ".venv",
    "__pycache__",
    "*.egg-info",
]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # do not perform function calls in argument defaults
    "B904",   # raise without from inside except
]

[tool.ruff.lint.isort]
known-first-party = ["src"]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["ARG001", "ARG002"]  # Unused arguments OK in tests

# =============================================================================
# Black configuration (T004) - Formatting
# =============================================================================
[tool.black]
target-version = ["py311"]
line-length = 100
exclude = '''
/(
    \.git
    | \.venv
    | __pycache__
    | \.egg-info
)/
'''

# =============================================================================
# MyPy configuration
# =============================================================================
[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
plugins = ["pydantic.mypy"]

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false
